+++ api/src/main/java/io/opencensus/tags/NoopTags.java
@@ -19,6 +19,7 @@ package io.opencensus.tags;
 import io.opencensus.common.Scope;
 import io.opencensus.internal.NoopScope;
 import io.opencensus.internal.Utils;
+import io.opencensus.tags.TagContextBuilder.TagScope;
 import io.opencensus.tags.propagation.TagContextBinarySerializer;
 import io.opencensus.tags.propagation.TagPropagationComponent;
 import java.util.Collections;
@@ -163,6 +164,12 @@ final class NoopTags {
     }
 
     @Override
+    public TagContextBuilder setTagScope(TagScope tagScope) {
+      Utils.checkNotNull(tagScope, "tagScope");
+      return null;
+    }
+
+    @Override
     public TagContext build() {
       return getNoopTagContext();
     }
@@ -181,6 +188,11 @@ final class NoopTags {
     @Override
     protected Iterator<Tag> getIterator() {
       return Collections.<Tag>emptySet().iterator();
+    }
+
+    @Override
+    public TagScope getTagScope() {
+      return TagScope.LOCAL;
     }
   }