+++ api/src/test/java/io/opencensus/tags/InternalUtilsTest.java
@@ -19,6 +19,7 @@ package io.opencensus.tags;
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.collect.Lists;
+import io.opencensus.tags.TagContextBuilder.TagScope;
 import java.util.Iterator;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -37,6 +38,11 @@ public final class InternalUtilsTest {
           @Override
           protected Iterator<Tag> getIterator() {
             return iterator;
+          }
+
+          @Override
+          public TagScope getTagScope() {
+            return TagScope.LOCAL;
           }
         };
     assertThat(InternalUtils.getTags(ctx)).isSameAs(iterator);