+++ impl_core/src/main/java/io/opencensus/implcore/tags/TagMapImpl.java
@@ -18,6 +18,7 @@ package io.opencensus.implcore.tags;
 
 import io.opencensus.tags.Tag;
 import io.opencensus.tags.TagContext;
+import io.opencensus.tags.TagContextBuilder.TagScope;
 import io.opencensus.tags.TagKey;
 import io.opencensus.tags.TagValue;
 import java.util.Collections;
@@ -31,17 +32,26 @@ import javax.annotation.concurrent.Immutable;
 @Immutable
 public final class TagMapImpl extends TagContext {
 
-  public static final TagMapImpl EMPTY = new TagMapImpl(Collections.<TagKey, TagValue>emptyMap());
+  public static final TagMapImpl EMPTY =
+      new TagMapImpl(Collections.<TagKey, TagValue>emptyMap(), TagScope.LOCAL);
 
   // The types of the TagKey and value must match for each entry.
   private final Map<TagKey, TagValue> tags;
 
-  public TagMapImpl(Map<? extends TagKey, ? extends TagValue> tags) {
+  private final TagScope tagScope;
+
+  public TagMapImpl(Map<? extends TagKey, ? extends TagValue> tags, TagScope tagScope) {
     this.tags = Collections.unmodifiableMap(new HashMap<TagKey, TagValue>(tags));
+    this.tagScope = tagScope;
   }
 
   public Map<TagKey, TagValue> getTags() {
     return tags;
+  }
+
+  @Override
+  public TagScope getTagScope() {
+    return tagScope;
   }
 
   @Override