+++ impl_core/src/main/java/io/opencensus/implcore/tags/propagation/TagContextBinarySerializerImpl.java
@@ -20,6 +20,7 @@ import io.opencensus.implcore.internal.CurrentState;
 import io.opencensus.implcore.internal.CurrentState.State;
 import io.opencensus.implcore.tags.TagMapImpl;
 import io.opencensus.tags.TagContext;
+import io.opencensus.tags.TagContextBuilder.TagScope;
 import io.opencensus.tags.propagation.TagContextBinarySerializer;
 import io.opencensus.tags.propagation.TagContextDeserializationException;
 import io.opencensus.tags.propagation.TagContextSerializationException;
@@ -35,7 +36,7 @@ final class TagContextBinarySerializerImpl extends TagContextBinarySerializer {
 
   @Override
   public byte[] toByteArray(TagContext tags) throws TagContextSerializationException {
-    return state.getInternal() == State.DISABLED
+    return state.getInternal() == State.DISABLED || TagScope.LOCAL.equals(tags.getTagScope())
         ? EMPTY_BYTE_ARRAY
         : SerializationUtils.serializeBinary(tags);
   }