+++ impl_core/src/test/java/io/opencensus/implcore/stats/StatsTestUtil.java
@@ -41,6 +41,7 @@ import io.opencensus.stats.ViewData.AggregationWindowData.CumulativeData;
 import io.opencensus.stats.ViewData.AggregationWindowData.IntervalData;
 import io.opencensus.tags.Tag;
 import io.opencensus.tags.TagContext;
+import io.opencensus.tags.TagContextBuilder.TagScope;
 import io.opencensus.tags.TagValue;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -218,6 +219,11 @@ final class StatsTestUtil {
     @Override
     protected Iterator<Tag> getIterator() {
       return tags.iterator();
+    }
+
+    @Override
+    public TagScope getTagScope() {
+      return TagScope.LOCAL;
     }
   }
 }