+++ impl_core/src/test/java/io/opencensus/implcore/tags/TagMapImplTest.java
@@ -28,6 +28,7 @@ import io.opencensus.implcore.internal.CurrentState.State;
 import io.opencensus.tags.Tag;
 import io.opencensus.tags.TagContext;
 import io.opencensus.tags.TagContextBuilder;
+import io.opencensus.tags.TagContextBuilder.TagScope;
 import io.opencensus.tags.TagKey;
 import io.opencensus.tags.TagValue;
 import io.opencensus.tags.Tagger;
@@ -43,7 +44,7 @@ import org.junit.runners.JUnit4;
 /**
  * Tests for {@link TagMapImpl} and {@link TagMapBuilderImpl}.
  *
- * <p>Tests for {@link TagMapBuilderImpl#buildScoped()} are in {@link ScopedTagContextsTest}.
+ * <p>Tests for {@link TagMapBuilderImpl#buildScoped()} are in {@link ScopedTagMapTest}.
  */
 @RunWith(JUnit4.class)
 public class TagMapImplTest {
@@ -59,33 +60,33 @@ public class TagMapImplTest {
 
   @Test
   public void getTags_empty() {
-    TagMapImpl tags = new TagMapImpl(ImmutableMap.<TagKey, TagValue>of());
+    TagMapImpl tags = new TagMapImpl(ImmutableMap.<TagKey, TagValue>of(), TagScope.LOCAL);
     assertThat(tags.getTags()).isEmpty();
   }
 
   @Test
   public void getTags_nonEmpty() {
-    TagMapImpl tags = new TagMapImpl(ImmutableMap.of(K1, V1, K2, V2));
+    TagMapImpl tags = new TagMapImpl(ImmutableMap.of(K1, V1, K2, V2), TagScope.LOCAL);
     assertThat(tags.getTags()).containsExactly(K1, V1, K2, V2);
   }
 
   @Test
   public void put_newKey() {
-    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1));
+    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1), TagScope.LOCAL);
     assertThat(((TagMapImpl) tagger.toBuilder(tags).put(K2, V2).build()).getTags())
         .containsExactly(K1, V1, K2, V2);
   }
 
   @Test
   public void put_existingKey() {
-    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1));
+    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1), TagScope.LOCAL);
     assertThat(((TagMapImpl) tagger.toBuilder(tags).put(K1, V2).build()).getTags())
         .containsExactly(K1, V2);
   }
 
   @Test
   public void put_nullKey() {
-    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1));
+    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1), TagScope.LOCAL);
     TagContextBuilder builder = tagger.toBuilder(tags);
     thrown.expect(NullPointerException.class);
     thrown.expectMessage("key");
@@ -94,7 +95,7 @@ public class TagMapImplTest {
 
   @Test
   public void put_nullValue() {
-    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1));
+    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1), TagScope.LOCAL);
     TagContextBuilder builder = tagger.toBuilder(tags);
     thrown.expect(NullPointerException.class);
     thrown.expectMessage("value");
@@ -103,21 +104,21 @@ public class TagMapImplTest {
 
   @Test
   public void remove_existingKey() {
-    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1, K2, V2));
+    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1, K2, V2), TagScope.LOCAL);
     assertThat(((TagMapImpl) tagger.toBuilder(tags).remove(K1).build()).getTags())
         .containsExactly(K2, V2);
   }
 
   @Test
   public void remove_differentKey() {
-    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1));
+    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1), TagScope.LOCAL);
     assertThat(((TagMapImpl) tagger.toBuilder(tags).remove(K2).build()).getTags())
         .containsExactly(K1, V1);
   }
 
   @Test
   public void remove_nullKey() {
-    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1));
+    TagContext tags = new TagMapImpl(ImmutableMap.of(K1, V1), TagScope.LOCAL);
     TagContextBuilder builder = tagger.toBuilder(tags);
     thrown.expect(NullPointerException.class);
     thrown.expectMessage("key");
@@ -126,7 +127,7 @@ public class TagMapImplTest {
 
   @Test
   public void testIterator() {
-    TagMapImpl tags = new TagMapImpl(ImmutableMap.of(K1, V1, K2, V2));
+    TagMapImpl tags = new TagMapImpl(ImmutableMap.of(K1, V1, K2, V2), TagScope.LOCAL);
     Iterator<Tag> i = tags.getIterator();
     assertTrue(i.hasNext());
     Tag tag1 = i.next();
@@ -140,11 +141,17 @@ public class TagMapImplTest {
 
   @Test
   public void disallowCallingRemoveOnIterator() {
-    TagMapImpl tags = new TagMapImpl(ImmutableMap.of(K1, V1, K2, V2));
+    TagMapImpl tags = new TagMapImpl(ImmutableMap.of(K1, V1, K2, V2), TagScope.LOCAL);
     Iterator<Tag> i = tags.getIterator();
     i.next();
     thrown.expect(UnsupportedOperationException.class);
     i.remove();
+  }
+
+  @Test
+  public void getTagScope() {
+    TagMapImpl tags = new TagMapImpl(ImmutableMap.<TagKey, TagValue>of(), TagScope.REQUEST);
+    assertThat(tags.getTagScope()).isEqualTo(TagScope.REQUEST);
   }
 
   @Test
@@ -158,6 +165,11 @@ public class TagMapImplTest {
               @Override
               protected Iterator<Tag> getIterator() {
                 return Lists.<Tag>newArrayList(Tag.create(K1, V1), Tag.create(K2, V2)).iterator();
+              }
+
+              @Override
+              public TagScope getTagScope() {
+                return TagScope.LOCAL;
               }
             })
         .addEqualityGroup(tagger.emptyBuilder().put(K1, V1).put(K2, V1).build())